requirements.txt
exif==1.6.0
geographiclib==2.0
geopy==2.4.1
plum-py==0.8.7
requests==2.32.3
pillow==10.4.0
argparse==1.4.0
numpy==1.26.4
openCV-python==4.5.3
cryptography==36.0.0
matplotlib==3.5.1